Refactors for New Labels
The feature support and overall feature variance plots now includes bars for the new primal problems.
Other important refactors: `features_to_ablate` is now programmatically selected by choosing features with mean contributions less than +/- 1e-3.
We continue to be thwarted by the problem of verifying whether performance is actually as high as it is reported to be, for now.