Feature selection: Use agreement between three test statistics
Now, instead of ablating many features at once, we test each feature stepwise based on the chi-squared test for independence, the mutual information between feature and target, and the F test for group difference between feature and target. If these three methods agree that a feature is most informative, we choose to keep it in the subset.